fix: consolidate pkg/config files to satisfy pkg-file-count limit
Merge load/persist helpers into existing config modules and relocate
blocking-load validation tests to pkg/config/load so CI pkg-file-count
passes without changing load/persist behavior.
