+++ pkg/config/config_validator_topology.go
@@ -806,4 +806,19 @@ func ruleCanonicalStructuralValidation(cfg *interfaces.FactoryConfig) []Finding 
 	return CanonicalStructuralFindings(cfg)
 }
 
+func validateBlockingFactoryLoad(cfg *interfaces.FactoryConfig) error {
+	if cfg == nil {
+		return nil
+	}
+	result := validation.ValidateBlockingLoad(cfg)
+	if len(result.Targets) == 0 {
+		return nil
+	}
+	return fmt.Errorf(
+		"%w: factory topology contains invalid graph references (%d blocking validation targets)",
+		ErrInvalidNamedFactory,
+		len(result.Targets),
+	)
+}
+
 // --- Rule: resource usage validation ---
